pyproject.toml
[project]
name = "cursor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "aioimaplib>=2.0.1",
    "aiosqlite>=0.21.0",
    "asyncio>=3.4.3",
    "curl-cffi>=0.9.0",
    "dataclasses>=0.8",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pycryptodome>=3.21.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "typing>=3.10.0.0",
]
